Move ControlReference and ExpressionParser files under ControllerEmu
