+++ Source/Core/InputCommon/CMakeLists.txt
@@ -14,12 +14,12 @@ set(SRCS InputConfig.cpp
   ControllerEmu/ControlGroup/Slider.cpp
   ControllerEmu/ControlGroup/Tilt.cpp
   ControllerEmu/ControlGroup/Triggers.cpp
+  ControllerEmu/ControlReference/ControlReference.cpp
+  ControllerEmu/ControlReference/ExpressionParser.cpp
   ControllerEmu/Setting/BooleanSetting.cpp
   ControllerEmu/Setting/NumericSetting.cpp
   ControllerInterface/ControllerInterface.cpp
   ControllerInterface/Device.cpp
-  ControlReference/ControlReference.cpp
-  ControlReference/ExpressionParser.cpp
 )
 set(LIBS common)
