+++ Source/Core/InputCommon/ControllerEmu/Control/Control.cpp
@@ -5,7 +5,7 @@
 #include "InputCommon/ControllerEmu/Control/Control.h"
 
 #include <utility>
-#include "InputCommon/ControlReference/ControlReference.h"
+#include "InputCommon/ControllerEmu/ControlReference/ControlReference.h"
 
 namespace ControllerEmu
 {
