+++ Source/Core/InputCommon/ControllerEmu/ControlGroup/Tilt.cpp
@@ -10,9 +10,9 @@
 #include <string>
 
 #include "Common/Common.h"
-#include "InputCommon/ControlReference/ControlReference.h"
 #include "InputCommon/ControllerEmu/Control/Control.h"
 #include "InputCommon/ControllerEmu/Control/Input.h"
+#include "InputCommon/ControllerEmu/ControlReference/ControlReference.h"
 #include "InputCommon/ControllerEmu/Setting/NumericSetting.h"
 
 namespace ControllerEmu
