+++ Source/Core/InputCommon/ControllerEmu/ControlGroup/Triggers.cpp
@@ -10,8 +10,8 @@
 #include <string>
 
 #include "Common/Common.h"
-#include "InputCommon/ControlReference/ControlReference.h"
 #include "InputCommon/ControllerEmu/Control/Control.h"
+#include "InputCommon/ControllerEmu/ControlReference/ControlReference.h"
 #include "InputCommon/ControllerEmu/Setting/NumericSetting.h"
 
 namespace ControllerEmu
