+++ Source/Core/InputCommon/ControllerEmu/ControlReference/ExpressionParser.cpp
@@ -11,7 +11,7 @@
 #include <vector>
 
 #include "Common/StringUtil.h"
-#include "InputCommon/ControlReference/ExpressionParser.h"
+#include "InputCommon/ControllerEmu/ControlReference/ExpressionParser.h"
 
 using namespace ciface::Core;
