+++ Source/Core/InputCommon/InputCommon.vcxproj
@@ -51,6 +51,8 @@
     <ClCompile Include="ControllerEmu\ControlGroup\Slider.cpp" />
     <ClCompile Include="ControllerEmu\ControlGroup\Tilt.cpp" />
     <ClCompile Include="ControllerEmu\ControlGroup\Triggers.cpp" />
+    <ClCompile Include="ControllerEmu\ControlReference\ControlReference.cpp" />
+    <ClCompile Include="ControllerEmu\ControlReference\ExpressionParser.cpp" />
     <ClCompile Include="ControllerEmu\Setting\BooleanSetting.cpp" />
     <ClCompile Include="ControllerEmu\Setting\NumericSetting.cpp" />
     <ClCompile Include="ControllerInterface\ControllerInterface.cpp" />
@@ -59,8 +61,6 @@
     <ClCompile Include="ControllerInterface\DInput\DInputJoystick.cpp" />
     <ClCompile Include="ControllerInterface\DInput\DInputKeyboardMouse.cpp" />
     <ClCompile Include="ControllerInterface\DInput\XInputFilter.cpp" />
-    <ClCompile Include="ControlReference\ControlReference.cpp" />
-    <ClCompile Include="ControlReference\ExpressionParser.cpp" />
     <ClCompile Include="ControllerInterface\ForceFeedback\ForceFeedbackDevice.cpp" />
     <ClCompile Include="ControllerInterface\XInput\XInput.cpp" />
     <ClCompile Include="GCAdapter.cpp">
@@ -91,6 +91,8 @@
     <ClInclude Include="ControllerEmu\Setting\BooleanSetting.h" />
     <ClInclude Include="ControllerEmu\Setting\NumericSetting.h" />
     <ClInclude Include="ControllerEmu\Setting\Setting.h" />
+    <ClInclude Include="ControllerEmu\ControlReference\ControlReference.h" />
+    <ClInclude Include="ControllerEmu\ControlReference\ExpressionParser.h" />
     <ClInclude Include="ControllerInterface\ControllerInterface.h" />
     <ClInclude Include="ControllerInterface\Device.h" />
     <ClInclude Include="ControllerInterface\DInput\DInput.h" />
@@ -98,8 +100,6 @@
     <ClInclude Include="ControllerInterface\DInput\DInputJoystick.h" />
     <ClInclude Include="ControllerInterface\DInput\DInputKeyboardMouse.h" />
     <ClInclude Include="ControllerInterface\DInput\XInputFilter.h" />
-    <ClInclude Include="ControlReference\ControlReference.h" />
-    <ClInclude Include="ControlReference\ExpressionParser.h" />
     <ClInclude Include="ControllerInterface\ForceFeedback\ForceFeedbackDevice.h" />
     <ClInclude Include="ControllerInterface\XInput\XInput.h" />
     <ClInclude Include="GCAdapter.h" />
